feat: separa funcoes em modulos
